Address compilation failure with Clang -std=c++20. (#56)
In C++20, vector became constexpr. Somehow, this causes the calls to
`sub_nodes.emplace_back` to trigger an error from the attempt to
convert `unique_ptr<MatmulNode>` (and others) to `unique_ptr<INode>`,
when `MatmulNode` is an incomplete type.

It seems like that _should_ have been an error before as well, since
that conversion is only valid if `MatmulNode` is known to be a subtype
of INode. But, I suppose previously it was deferred until after
`matmul.h` had been included, and now due to constexpr the conversion
attempt is triggered eagerly.

To resolve, put the definitions of the functions in the headers
associated with each type.

Co-authored-by: Anerudhan Gopal <agopal@nvidia.com>