CMake improvements for third party users
This commit aims to improve the use of cudnn-frontend as a third party
package in CMake projects, and is a result of my efforts of trying to
depend on cudnn-frontend in a Nix-managed ecosystem. It also adds
cudnn to the header library, rather than having dependant projects
add cudnn themselves.

With this commit, users should be able to simply use:

```
find_package(cudnn_frontend REQUIRED)

add_executable(a-dnn-project src/main.cpp)
target_link_libraries(a-dnn-project
                      PRIVATE
                      cudnn_frontend::cudnn_frontend)
```

without requiring additional find_package and target_link_libraries for
libraries that cudnn_frontend depends on.

Some modification was also made to testing, so that a Nix `checkPhase`
will run the tests. Unfortunately the tests are not passing, but I
suspect this is expected.